fix iso8601ToSecond converter and add test
